Refactor LogInitalization to a common trait
