+++ .travis.yml
@@ -7,5 +7,5 @@ before_script:
   - sudo apt-get install -y rpm
 script:
   #- sbt ++$TRAVIS_SCALA_VERSION test
-  - sbt ++$TRAVIS_SCALA_VERSION seqexec/universal:packageZipTarball
-  - sbt ++$TRAVIS_SCALA_VERSION seqexec/rpm:packageBin
+  - sbt ++$TRAVIS_SCALA_VERSION seqexec_server/universal:packageZipTarball
+  #- sbt ++$TRAVIS_SCALA_VERSION seqexec/rpm:packageBin
